+++ font_rasterizer_example/src/lib.rs
@@ -8,10 +8,11 @@ use cgmath::Rotation3;
 use font_rasterizer::{
     color_theme::ColorTheme,
     context::{StateContext, WindowSize},
-    instances::{GlyphInstances, InstanceAttributes},
+    instances::GlyphInstances,
     motion::{EasingFuncType, MotionDetail, MotionFlags, MotionTarget, MotionType},
     rasterizer_pipeline::Quarity,
     time::now_millis,
+    vector_instances::InstanceAttributes,
 };
 use log::info;
 use ui_support::{
