+++ ui_support/examples/save_apng.rs
@@ -6,10 +6,11 @@ use cgmath::Rotation3;
 use font_rasterizer::{
     color_theme::ColorTheme,
     context::{StateContext, WindowSize},
-    instances::{GlyphInstances, InstanceAttributes},
+    instances::GlyphInstances,
     motion::{EasingFuncType, MotionDetail, MotionFlags, MotionTarget, MotionType},
     rasterizer_pipeline::Quarity,
     time::now_millis,
+    vector_instances::InstanceAttributes,
 };
 use log::{debug, info};
 use ui_support::{
