+++ ui_support/src/ui/single_line.rs
@@ -4,11 +4,8 @@ use cgmath::Rotation3;
 use instant::Duration;
 
 use font_rasterizer::{
-    char_width_calcurator::CharWidthCalculator,
-    context::StateContext,
-    instances::{GlyphInstances, InstanceAttributes},
-    motion::MotionFlags,
-    time::now_millis,
+    char_width_calcurator::CharWidthCalculator, context::StateContext, instances::GlyphInstances,
+    motion::MotionFlags, time::now_millis, vector_instances::InstanceAttributes,
 };
 
 use super::get_color;
